refactor(sdk)!: delete the engine, leaving one ACP stack
Nothing in `apps/bitrouter` has driven `engine::Session` since `chat` moved
to the shared client. This removes it and everything that existed only to
serve it: the ACP `Pipeline` and its three hook traits, `executor`,
`turn::TurnController`, `permissions::PermissionRegistry`,
`session::SessionState`, `config_routing::ConfigAcpRoutingTable`,
`telemetry::TelemetryHook`, and `up::UpstreamConnection`.

The pipeline is worth naming: it routed through a table pinned to one
target that its executor then ignored, and no consumer outside this crate
ever registered a hook on it. Its one live output was a per-turn record,
and latency and stop reason turned out to be visible on the prompt
round-trip all along — which is where `report_turn` reads them now.

Two types outlive their module because the app still needs them, and
neither belongs to any surviving SDK type: `LaunchOptions`, whose three
fields are now consumed by three different places, and `RequestCompleted`,
which is what the span recorder and the per-turn log take. Both move to
`acp_cli`. `ContextUsage` stays in the SDK — it is what ACP actually
carries.

`acp/mod.rs` also loses two `pub use` re-exports that violated the house
rule against re-exporting inside a public mod; consumers now name
`acp::transport::` directly.

Test count 2910 → 2881. Every one of the 30 belonged to deleted code or to
an invariant the amendment deliberately drops (I9's turn queue, and
reattach replay). The two that pin **I1** and **I10** live in `up.rs` and
were written against `UpstreamConnection`; they are ported onto
`AgentProcess` + `AcpClient` rather than deleted with the type they
happened to be written on, since the child they drive is unchanged.
