+++ apps/bitrouter/tests/acp.rs
@@ -16,7 +16,7 @@
 
 use std::collections::HashMap;
 
-use bitrouter_sdk::acp::{AcpAgentConfig, AcpTransport};
+use bitrouter_sdk::acp::transport::{AcpAgentConfig, AcpTransport};
 use bitrouter_sdk::config::Config;
 
 /// Bash ACP stub: initialize → session/new → prompt emits one update then
